De-nest RowTest
This actually runs the inner tests
